refactor (OverviewItem): extract helpers to utils
